requirements.txt
cython
dj-static==0.0.6
django-autoslug
django-grappelli
django-typed-models
Django>=1.10,<2.0
djangorestframework
isodate
jsonfield
matplotlib>2.0,<3.0
netCDF4
numpy
pandas
pyaxiom
pyproj
pysgrid
python-dateutil
pytz
pyugrid
rtree
utide
cmocean

celery>=4.1
hiredis
django-redis-cache
